[COR-891] clean up and fixing a bug
