+++ http/context.go
@@ -8,8 +8,6 @@ import (
 	"mime"
 	"net/http"
 	"net/url"
-	"reflect"
-	"strconv"
 	"strings"
 
 	"github.com/beaconsoftwarellc/gadget/v2/errors"
@@ -17,6 +15,7 @@ import (
 	"github.com/beaconsoftwarellc/gadget/v2/stringutil"
 	qerror "github.com/beaconsoftwarellc/quimby/v2/error"
 	"github.com/beaconsoftwarellc/quimby/v2/http/multipartform"
+	"github.com/beaconsoftwarellc/quimby/v2/http/urlencodedform"
 )
 
 // NoContentError is returned when Read is called and the Request has a 0
@@ -196,107 +195,6 @@ func (context *Context) readJSON(body []byte, target interface{}) error {
 	return json.Unmarshal(body, target)
 }
 
-// readForm takes the form-urlencoded content type body of the Request and populates an object
-// of the same type as the passed interface{}
-func (context *Context) readForm(body []byte, target interface{}) error {
-	values, err := url.ParseQuery(string(body))
-	if nil != err {
-		return err
-	}
-	return context.valuesToObject(values, target)
-}
-
-func (context *Context) inspectModel(target interface{}) map[string]reflect.Type {
-	v := reflect.Indirect(reflect.ValueOf(target))
-	fieldMap := make(map[string]reflect.Type, v.NumField())
-	for i := 0; i < v.NumField(); i++ {
-		fieldMap[v.Type().Field(i).Name] = v.Field(i).Type()
-	}
-	return fieldMap
-}
-
-func (context *Context) valuesToArray(fieldType reflect.Type, values []string) []interface{} {
-	elemKind := fieldType.Elem().Kind()
-	anon := make([]interface{}, len(values))
-	coerce := func(v string) interface{} { return v }
-	if elemKind == reflect.Int {
-		coerce = func(v string) interface{} {
-			i, err := strconv.Atoi(v)
-			if nil != err {
-				log.Warnf("error encountered coercing url value '%s' to int: %s", v, err)
-			}
-			return i
-		}
-	} else if elemKind == reflect.Bool {
-		coerce = func(v string) interface{} {
-			return v == "true"
-		}
-	} else if elemKind != reflect.String {
-		log.Warnf("valuesToArray recieved unhandled kind %v", elemKind)
-		return nil
-	}
-	for i, v := range values {
-		anon[i] = coerce(v)
-	}
-	return anon
-}
-
-func breakUpArrayValuedParameter(values []string) []string {
-	newValues := make([]string, 0, len(values))
-	for i := 0; i < len(values); i++ {
-		newValues = append(newValues, strings.Split(values[i], ",")...)
-	}
-	return newValues
-}
-
-func (context *Context) valuesToObject(values url.Values, target interface{}) error {
-	var err error
-	fieldMap := context.inspectModel(target)
-	valueMap := make(map[string]interface{})
-	for fieldName, fieldType := range fieldMap {
-		urlFieldName := stringutil.Underscore(fieldName)
-		queryValues := values[urlFieldName]
-		queryValues = append(queryValues,
-			breakUpArrayValuedParameter(values[urlFieldName+"[]"])...)
-
-		if len(queryValues) == 0 {
-			continue
-		}
-		switch fieldType.Kind() {
-		case reflect.Chan, reflect.Func, reflect.Interface, reflect.Map, reflect.Struct, reflect.UnsafePointer:
-			continue
-		case reflect.Slice, reflect.Array:
-			if len(queryValues) == 0 {
-				continue
-			}
-			if arrayValues := context.valuesToArray(fieldType, queryValues); nil != arrayValues {
-				valueMap[fieldName] = arrayValues
-			} else {
-				log.Warnf("failed to assigned values to array for field '%s' kind '%s'",
-					urlFieldName, fieldType.Kind())
-			}
-		case reflect.String:
-			valueMap[fieldName] = queryValues[0]
-		case reflect.Int:
-			valueMap[fieldName], err = strconv.Atoi(queryValues[0])
-			if nil != err {
-				log.Warnf("error parsing int for field name '%s' and value '%s'",
-					fieldName, queryValues[0])
-			}
-		case reflect.Bool:
-			valueMap[fieldName] = queryValues[0] == "true"
-
-		default:
-			log.Warnf("unhandled kind %v will be omitted from url values", fieldType)
-		}
-	}
-	data, err := json.Marshal(valueMap)
-	if nil != err {
-		return err
-	}
-	return json.Unmarshal(data, target)
-}
-
 // ReadObject reads the body of the Request and unmarshals an object the
 // same type as the passed implementation of interface{}
 func (context *Context) ReadObject(target interface{}) error {
@@ -313,13 +211,13 @@ func (context *Context) ReadObject(target interface{}) error {
 	}
 	switch contentType {
 	case contentTypeForm:
-		err = context.readForm(body, target)
+		err = urlencodedform.Unmarshal(body, target)
 	case contentTypeMultiPartFormData:
 		fallthrough
 	case contentTypeMultiPartFormData1:
 		err = context.Request.
 			ParseMultipartForm(multipartform.MaxMemoryBytes)
-		if nil != err {
+		if nil == err {
 			err = multipartform.Unmarshal(context.Request.MultipartForm, target)
 		}
 	case contentTypeJSON:
@@ -352,5 +250,5 @@ func (context *Context) Write(bytes []byte) error {
 
 // ReadQueryParams converts URL Parameters into an Object
 func (context *Context) ReadQueryParams(target interface{}) error {
-	return context.valuesToObject(context.URLParameters, target)
+	return urlencodedform.URLValuesToObject(context.URLParameters, target)
 }
